requirements.txt
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
opencv-python==4.9.0.80
deepface==0.0.79
tensorflow==2.16.1
numpy==1.26.4
Pillow==10.2.0
celery==5.3.6
redis==5.0.4
